manda_project/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
coverage==7.2.7
cryptography==41.0.5
daphne==4.0.0
Django==3.2.22
django-cors-headers==4.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.7.0
incremental==22.10.0
inflection==0.5.1
msgpack==1.0.5
packaging==23.2
Pillow==9.5.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
uritemplate==4.1.1
zipp==3.15.0
zope.interface==6.1